add tests for postgres (#1)
The `schema` column is currently of type JSONB ~which is not supported by sqlite~ (see comment below). Here we add an integration test that creates and tears down a postgres instance for testing purposes. I'm open to shredding this we think it's a bad idea (it did help me start onboarding to the project).

Pros:
* Use postgres for testing and production;
* Assign it a "test" profile in docker-compose so that it is otherwise ignored.

Cons:
* Adds complexity to integration tests;
* Some hacks (e.g., "wait 3 seconds for postgres to start");
* Tearing down test DB calls `DROP SCHEMA public CASCADE`; a careless error could wipe the real DB.

I copy/pasted the existing (broken I think) unit test and updated it. Marking WIP as if we do want to merge this, we should delete that file.