add brute force iteration and deduping (#23)
Here we change the default behavior of
`extract_using_existing_extractor` to chunk up the text, extract over
chunks and consolidate the results.

We add constants `MODEL_NAME` and `CHUNK_SIZE`, with the intent of
chunking based on the token window length of the model.

Importantly: since we are now running extraction more than once, we need
a strategy for consolidating the results. Here we merge them into a list
and remove strict duplicates. Looking for feedback on this.